Move pending Canvas migration to version 2.40
